GPT/requirements.txt
fire>=0.1.3
regex==2017.4.5
requests==2.21.0
tqdm==4.31.1
toposort==1.5
tensorflow-gpu==1.15
numpy==1.14.3


